upgrade ember-composable-helpers to avoid array helper error
When upgrading labs-zap-search to ember octane 3.17, there arose an
error with conflicting helpers (the ember array helper we us in
template and the array helper associated with ember-composable-helpers).
In order to avoid this error, we needed to upgrade the newer version
of ember-composable-helpers, where the "array" helper is not included
in the node-modules.